requirements.txt
appdirs==1.4.3
argon2-cffi==19.2.0
asgiref==3.2.3
astroid==2.3.3
Babel==2.8.0
beautifulsoup4==4.6.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
distlib==0.3.0
Django==3.0.3
django-appconf==1.0.3
django-axes==5.3.1
django-crispy-forms==1.8.1
django-formtools==2.2
django-ipware==2.1.0
django-lockout==0.1.1
django-modelcluster==5.0.1
django-otp==0.8.1
django-phonenumber-field==3.0.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
django-taggit==1.2.0
django-treebeard==4.3.1
django-two-factor-auth==1.10.0
djangorestframework==3.11.0
draftjs-exporter==2.1.7
filelock==3.0.12
html5lib==1.0.1
idna==2.9
importlib-metadata==1.5.0
isort==4.3.21
l18n==2018.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==6.2.2
pycparser==2.20
pylint==2.4.4
pytz==2019.3
qrcode==6.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
Unidecode==1.1.1
urllib3==1.25.8
virtualenv==20.0.9
wagtail==2.8
webencodings==0.5.1
Willow==1.3
wrapt==1.11.2
zipp==3.1.0
